Bugfix: "git pull" cannot be made on a tagged version of the rtems kernel.
In git the philosophy is that you cannot do "git pull" when you did
"git checkout [TAG]" before. "git pull" only works on checked out branches,
not tags.

In rtems/config/tools/rtems-kernel-4-1.cfg "git pull" was removed. A specified
version of the RTEMS kernel should be exactly that. If we want to have the
*newest* RTEMS kernel of a branch we should do this with a new *.bset file.