Bugfix: A misconfiguration could update the wrong git working copy.
A function in git.py checked the validity of a git repository by testing
whether the path of the working copy existed. If it did, it called "git
status" to see if the repository was valid. However, if the path existed but
had no ".git" directory in it, "git status" would look in directories further
up the directory hierarchy. If it found a git repository somewhere on the way
it would work on this repository.

In the end this could cause git.valid() returning True where it shouldn't.
Other parts of the source builder then executed "git reset --hard" on the
wrong local repository.

Now the program checks first if a directory ".git" in the given path exists,
and only then tries to execute "git status".